jhiRoot/plantsFront/src/main/webapp/app/entities/classificationMS/forme/route/forme-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IForme, Forme } from '../forme.model';
import { FormeService } from '../service/forme.service';

import { FormeRoutingResolveService } from './forme-routing-resolve.service';

describe('Forme routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: FormeRoutingResolveService;
  let service: FormeService;
  let resultForme: IForme | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(FormeRoutingResolveService);
    service = TestBed.inject(FormeService);
    resultForme = undefined;
  });

  describe('resolve', () => {
    it('should return IForme returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultForme = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultForme).toEqual({ id: 123 });
    });

    it('should return new IForme if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultForme = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultForme).toEqual(new Forme());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Forme })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultForme = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultForme).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
